Remove node_count from GameNode class
The "node_count" attribute in the GameNode class is removed. It was originally intended to track the number of nodes in the game tree but proved unnecessary and was not used in other operations. The GameNode's constructor and walk-back method have also been changed accordingly to remove their dependencies on "node_count".