Update iteration calculation in mcts.py
The number of iterations in run_mcts function is now calculated as the product of base_iter and number of legal moves. This ensures that as the action space increases, more iterations are taken for better exploration of the nodes. The rationale here is that the game is asymmetrical and different sides often have very different action-space sizes. If they are given the same number of MCTS iterations, the side with a bigger action-space will always end up with fewer visits per node and a shallower search, which disadvantages that side as an unintended by-product of how MCTS works. The change here fixes this by giving MCTS the number of base_iterations, which is the average visits per node we are targeting, and the algorithm dynamically calculates the total iterations that the player gets as base iterations * action space size.