The core file has been changed to a .pyx file, and the get_actions() method within the GameNode class has been Cythonized.
The get_actions method represented the majority of the computational expenses incurred in the game. Cythonizing it so that it is compiled to C resulted in a speedup of roughly 50%, enabling MCTS-based games with a large number of rollouts (10 rollouts per legal move) to be completed in 15 seconds on average.