Refactor dtype in core.pyx and add legal_actions attribute
The dtype of np.array has been changed from np.int8 to np.intc for better performance and compatibility. Additionally, a new attribute, 'legal_actions', has been added to the class in order to keep track of legal actions during the game. This is necessary to mask out the illegal actions in the policy prediction when training an agent.